refactor(indexer/handlers): extract fpmm/limits-and-fees.ts (TradingLimit + LiquidityStrategy + fee handlers) (PR-S4)
Final slice of the handlers/fpmm.ts split. Move TradingLimitConfigured,
LiquidityStrategyUpdated, the 3 fee handlers, and the shared
`updatePoolFeeField` helper into the new
indexer-envio/src/handlers/fpmm/limits-and-fees.ts (154 lines).

After this slice, handlers/fpmm.ts contains only FPMM.Swap (182 lines)
and is well below the 600-line soft cap. The four sibling modules each
own one event family:

- handlers/fpmm/factory.ts — pool creation + LP transfer
- handlers/fpmm/liquidity.ts — Mint + Burn
- handlers/fpmm/state-sync.ts — UpdateReserves + Rebalanced
- handlers/fpmm/limits-and-fees.ts — TradingLimits + Strategy + fees

EventHandlers.ts is updated to import the new module so Envio sees the
side-effect handler registrations at module load.

handlers/fpmm.ts: 322 → 182 lines.

Pure extraction. 380 tests pass.
