+++ indexer-envio/src/EventHandlers.ts
@@ -21,6 +21,7 @@ import "./handlers/fpmm";
 import "./handlers/fpmm/factory";
 import "./handlers/fpmm/liquidity";
 import "./handlers/fpmm/state-sync";
+import "./handlers/fpmm/limits-and-fees";
 import "./handlers/sortedOracles";
 import "./handlers/virtualPool";
 import "./handlers/feeToken";
